UI: Extract ComponentContainer
